Move AppInfo extension from Shared into Common (needed for BrowserKitInformation default initialization for strict concurrency).
